+++ TombEngine/Objects/Generic/Object/burning_torch.h
@@ -10,5 +10,4 @@ namespace TEN::Entities::Generic
 	void DoFlameTorch();
 	void GetFlameTorch();
 	void TorchControl(short itemNumber);
-	void FireCollision(short itemNumber, ItemInfo* laraItem, CollisionInfo* coll);
 }