+++ TombEngine/Objects/Generic/Switches/pulley_switch.cpp
@@ -3,6 +3,7 @@
 
 #include "Game/collision/collide_item.h"
 #include "Game/control/control.h"
+#include "Game/Hud/Hud.h"
 #include "Game/items.h"
 #include "Game/Lara/lara.h"
 #include "Game/Lara/lara_helpers.h"
@@ -12,6 +13,7 @@
 #include "Specific/Input/Input.h"
 #include "Specific/level.h"
 
+using namespace TEN::Hud;
 using namespace TEN::Input;
 
 enum PulleyFlags
@@ -87,6 +89,8 @@ namespace TEN::Entities::Switches
 	{
 		auto* laraInfo = GetLaraInfo(laraItem);
 		auto* switchItem = &g_Level.Items[itemNumber];
+
+		g_Hud.InteractionHighlighter.Test(*laraItem, *switchItem);
 
 		bool isUnderwater = (laraInfo->Control.WaterStatus == WaterStatus::Underwater);