+++ TombEngine/Objects/TR4/Object/StatuePlinth.cpp
@@ -8,6 +8,7 @@
 #include "Game/effects/effects.h"
 #include "Game/effects/tomb4fx.h"
 #include "Game/Gui.h"
+#include "Game/Hud/Hud.h"
 #include "Game/items.h"
 #include "Game/Lara/lara.h"
 #include "Game/Lara/lara_helpers.h"
@@ -18,6 +19,7 @@
 #include "Specific/level.h"
 
 using namespace TEN::Gui;
+using namespace TEN::Hud;
 using namespace TEN::Input;
 
 namespace TEN::Entities::TR4
@@ -49,7 +51,7 @@ namespace TEN::Entities::TR4
 		auto& keyHoleItem = g_Level.Items[itemNumber];
 		const auto& player = GetLaraInfo(playerItem);
 
-		short* triggerIndexPtr = GetTriggerIndex(&keyHoleItem);
+		g_Hud.InteractionHighlighter.Test(*playerItem, keyHoleItem);
 
 		if (!keyHoleItem.ItemFlags[0])
 			keyHoleItem.ItemFlags[0] = PLACE_PLINTHITEM_FRAME;
@@ -67,6 +69,7 @@ namespace TEN::Entities::TR4
 			keyItem = ID_PUZZLE_ITEM1;
 		}
 
+		short* triggerIndexPtr = GetTriggerIndex(&keyHoleItem);
 		if (triggerIndexPtr == nullptr)
 			return;