+++ TombEngine/Scripting/Internal/ReservedScriptNames.h
@@ -207,11 +207,7 @@ static constexpr char ScriptReserved_GetOnKilled[]			= "GetOnKilled";
 static constexpr char ScriptReserved_SetOnKilled[]			= "SetOnKilled";
 static constexpr char ScriptReserved_GetOnHit[]				= "GetOnHit";
 static constexpr char ScriptReserved_SetOnHit[]				= "SetOnHit";
-static constexpr char ScriptReserved_GetOnCollidedWithObject[]	= "GetOnCollidedWithObject";
-static constexpr char ScriptReserved_SetOnCollidedWithObject[]	= "SetOnCollidedWithObject";
-static constexpr char ScriptReserved_GetOnCollidedWithRoom[]	= "GetOnCollidedWithRoom";
-static constexpr char ScriptReserved_SetOnCollidedWithRoom[]	= "SetOnCollidedWithRoom";
-static constexpr char ScriptReserved_ToVec2[]					= "ToVec2";
+static constexpr char ScriptReserved_ToVec2[]				= "ToVec2";
 static constexpr char ScriptReserved_AttachObjCamera[]		= "AttachObjCamera";
 static constexpr char ScriptReserved_AnimFromObject[]		= "AnimFromObject";
 static constexpr char ScriptReserved_ClearActivators[]		= "ClearActivators";
@@ -219,6 +215,12 @@ static constexpr char ScriptReserved_IsMoveableInside[]		= "IsMoveableInside";
 static constexpr char ScriptReserved_GetFlag[]				= "GetFlag";
 static constexpr char ScriptReserved_SetFlag[]				= "SetFlag";
 static constexpr char ScriptReserved_IsTagPresent[]			= "IsTagPresent";
+
+static constexpr char ScriptReserved_GetOnCollidedWithObject[]	= "GetOnCollidedWithObject";
+static constexpr char ScriptReserved_SetOnCollidedWithObject[]	= "SetOnCollidedWithObject";
+static constexpr char ScriptReserved_GetOnCollidedWithRoom[]	= "GetOnCollidedWithRoom";
+static constexpr char ScriptReserved_SetOnCollidedWithRoom[]	= "SetOnCollidedWithRoom";
+static constexpr char ScriptReserved_ShowInteractionHighlight[] = "ShowInteractionHighlight";
 
 // Flow Functions
 static constexpr char ScriptReserved_AddLevel[]					= "AddLevel";