+++ CHANGELOG.md
@@ -8,6 +8,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 ## New features
 * Added animated textures support for all objects.
 * Added UVRotate support.
+* Added interaction highlighter.
 * Added bullet hole and explosion decals for room geometry.
 * Increased level loading speed by switching to LZ4 compression format.
 
@@ -27,6 +28,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed intense GPU usage when engine window is minimized.
 
 ### Lua API changes
+* Added `Moveable:ShowInteractionHighlight` method.
 * Renamed `ObjID.MESHSWAP_HITMAN` to `ObjID.MESHSWAP_CYBORG`.
 
 ## [Version 1.9.2](https://github.com/TombEngine/TombEditorReleases/releases/tag/v1.9.2) - 2025-08-17